broke enemies into their own files and directories
